Add --count to limit message-sends processed
Along with --quiet this allows drbdump to be used to sample a
distributed process.

There's a minor problem in that the response to the last message is not
processed, but there is no way of knowing when (or if) it will appear.
Due to this limitation `--count 500` will show 500 message-sends 499
responses.