Record per-message latency statistics
This shows how long a message takes to process, but the time includes
both network delays and remote processing overhead.  Messages taking a
long time to process (due to a large calculation, perhaps) will
naturally have high response times.

This commit indicates a great need for some sort of object to represent
the message-send.

This commit also indicates that the result summary should be broken up
and matched to each message individually as the latencies are.  This
will allow summing of the total allocation cost of each message.